requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
fastapi[all]
python-dotenv
